requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn>=0.30.0

# AI/ML APIs
openai>=1.0.0
pinecone-client>=4.0.0
google-generativeai>=0.3.0

# Web Scraping & Parsing
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Image Processing
pillow>=10.0.0

# Data Processing
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0

